Move folder loading into service
