+++ packages/web-app-files/src/views/PublicFiles.vue
@@ -66,10 +66,7 @@ import MixinMountSideBar from '../mixins/sidebar/mountSideBar'
 import { VisibilityObserver } from 'web-pkg/src/observer'
 import { ImageDimension, ImageType } from '../constants'
 import debounce from 'lodash-es/debounce'
-import omit from 'lodash-es/omit'
-import { buildResource } from '../helpers/resources'
 import { bus } from 'web-pkg/src/instance'
-import { useTask } from 'vue-concurrency'
 import { nextTick, computed, unref } from '@vue/composition-api'
 
 import ListLoader from '../components/FilesList/ListLoader.vue'
@@ -78,43 +75,11 @@ import NotFoundMessage from '../components/FilesList/NotFoundMessage.vue'
 import ListInfo from '../components/FilesList/ListInfo.vue'
 import Pagination from '../components/FilesList/Pagination.vue'
 import ContextActions from '../components/FilesList/ContextActions.vue'
-import { DavProperties, DavProperty } from 'web-pkg/src/constants'
-import { linkRoleUploaderFolder } from '../helpers/share'
-import { createLocationOperations, createLocationPublic } from '../router'
-
-// hacky, get rid asap, just a workaround
-const unauthenticatedUserReady = async (router, store) => {
-  // exit early which could happen if
-  // the resources get reloaded
-  // another application decided that the user is already provisioned
-  if (store.getters.userReady) {
-    return
-  }
-
-  // pretty low level, error prone and weak, add method to the store to obtain the publicToken
-  // it looks like that something was available in the past, store.state.Files.publicLinkInEdit ...
-  const publicToken = (router.currentRoute.params.item || '').split('/')[0]
-  const publicLinkPassword = store.getters['Files/publicLinkPassword']
-
-  await store.dispatch('loadCapabilities', {
-    publicToken,
-    ...(publicLinkPassword && { user: 'public', password: publicLinkPassword })
-  })
-
-  // ocis at the moment is not able to create archives for public links that are password protected
-  // till this is supported by the backend remove it hard as a workaround
-  // https://github.com/owncloud/web/issues/6515
-  if (publicLinkPassword) {
-    store.commit('SET_CAPABILITIES', {
-      capabilities: omit(store.getters.capabilities, ['files.archivers']),
-      version: store.getters.version
-    })
-  }
-
-  store.commit('SET_USER_READY', true)
-}
+import { createLocationOperations } from '../router'
+import { folderService } from '../services/folder'
 
 const visibilityObserver = new VisibilityObserver()
+
 export default {
   components: {
     ResourceTable,
@@ -157,56 +122,11 @@ export default {
       fileList.accentuateItem(payload.id)
     })
 
-    const loadResourcesTask = useTask(function* (signal, ref, sameRoute, path = null) {
-      ref.CLEAR_CURRENT_FILES_LIST()
-
-      try {
-        let resources = yield ref.$client.publicFiles.list(
-          path || ref.$route.params.item,
-          ref.publicLinkPassword,
-          DavProperties.PublicLink
-        )
-
-        // Redirect to files drop if the link has role "uploader"
-        const sharePermissions = parseInt(
-          resources[0].getProperty(DavProperty.PublicLinkPermission)
-        )
-        if (linkRoleUploaderFolder.bitmask(false) === sharePermissions) {
-          ref.$router.replace(
-            createLocationPublic('files-public-drop', {
-              params: { token: ref.$route.params.item }
-            })
-          )
-          return
-        }
-
-        resources = resources.map(buildResource)
-        ref.LOAD_FILES({ currentFolder: resources[0], files: resources.slice(1) })
-
-        refreshFileListHeaderPosition()
-      } catch (error) {
-        ref.SET_CURRENT_FOLDER(null)
-        console.error(error)
-
-        if (error.statusCode === 401) {
-          ref.redirectToResolvePage()
-          return
-        }
-      }
-
-      // this is a workAround till we have extended the bootProcess
-      // if a visitor is able to view the current page
-      // the user is ready and the TOO LATE provisioning can start.
-      // there is no other way at the moment to find out if:
-      // publicLink is password protected
-      // public link is viewable
-      // so we expect if the user is able to load resources, so he also is ready
-      yield unauthenticatedUserReady(ref.$router, ref.$store)
-      ref.accessibleBreadcrumb_focusAndAnnounceBreadcrumb(sameRoute)
-    }).restartable()
+    const loadResourcesTask = folderService.getTask()
 
     return {
       fileListHeaderY,
+      refreshFileListHeaderPosition,
       loadResourcesTask,
       paginatedResources,
       paginationPages,
